+++ CHANGELOG.md
@@ -3,10 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Pseudo-features for better handling and sorting of top-level multi-features.
+
 ### Fixed
 - Bug with non-protein coding genes and the `tag.mrna.primary` selector.
-- Bug with how the GFF3 writer handles multi-feature IDs, and some implications
-  for how top-level multi-features lacking children are printed and sorted.
+- Bug with how the GFF3 writer handles multi-feature IDs.
 
 ## [0.3.1] - 2017-02-02
 ### Fixed